Move role datastore to database directory
