+++ app/src/main/java/one/njk/celestidesk/di/DataStoreModule.kt
@@ -6,7 +6,7 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
-import one.njk.celestidesk.data.RolesDataStore
+import one.njk.celestidesk.database.RolesDataStore
 import javax.inject.Singleton
 
 @Module
